backend/requirements.txt
# ─────────────────────────────────────────────
# Core Django
# ─────────────────────────────────────────────
Django>=5.0,<6.0
django-cors-headers>=4.3.1

# ─────────────────────────────────────────────
# GraphQL
# ─────────────────────────────────────────────
graphene-django>=3.2.2
graphene>=3.3

# ─────────────────────────────────────────────
# Database — PostgreSQL (Neon)
# ─────────────────────────────────────────────
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# ─────────────────────────────────────────────
# Environment Variables
# ─────────────────────────────────────────────
python-dotenv>=1.0.1

# ─────────────────────────────────────────────
# Authentication — JWT
# ─────────────────────────────────────────────
PyJWT>=2.8.0
bcrypt>=4.1.2

# ─────────────────────────────────────────────
# URL Validation
# ─────────────────────────────────────────────
requests>=2.31.0          # HTTP reachability check
dnspython>=2.6.1          # DNS resolution check

# ─────────────────────────────────────────────
# QR Code Generation
# ─────────────────────────────────────────────
qrcode[pil]>=7.4.2
Pillow>=10.2.0

# ─────────────────────────────────────────────
# User Agent Parsing (Analytics)
# ─────────────────────────────────────────────
user-agents>=2.2.0

# ─────────────────────────────────────────────
# IP Geolocation (Analytics)
# ─────────────────────────────────────────────
geoip2>=4.8.0

# ─────────────────────────────────────────────
# AI Integration — Google Gemini (Optional)
# ─────────────────────────────────────────────
google-generativeai>=0.7.0

# ─────────────────────────────────────────────
# Static Files (Production)
# ─────────────────────────────────────────────
whitenoise>=6.6.0
